Put TestStateCleaner behind the `test-utils` feature flag
DAP_DEPLOYMENT no longer controls whether the TestStateCleaner is
enabled. Instead the `test-utils` feature is used to determine this in
order to be consistent with the daphne_server crate.
